feat: admin 앱을 FSD 구조로 재구축하고 라우팅 시스템 구현
앱 진입점을 app 레이어로 이동하고 모든 라우트를 연결하여 FSD 아키텍처 정규화를 완료합니다.

- App.tsx, main.tsx를 src/app 폴더로 이동하여 FSD 구조 준수
- index.html entry point 경로 수정 (/src/app/main.tsx)
- 라우팅 설정 확대 (FixedLayout 기반 6개 라우트 모두 연결)
- 라우트 경로를 ROUTES 상수 기반으로 정의하여 유지보수성 강화
